First set of changes trying to get the prompt to work on the farm.
Most of the changes are just file permissiosn, which was done to allow more than one person to make changes in the directory structure.

Substantive changes are in the summary.sh script where we've added diagnostic output to tell where the files are getting created (or not, as the case may be).